browser-automation/requirements.txt
browser-use>=0.1.37
google-generativeai>=0.4.0
python-dotenv>=0.19.0
playwright>=1.41.0
langchain-google-genai>=0.1.0 